Login code, unique entries, predictions
Created Login/Registration code, as well as made specific database entries unique to avoid duplicates. If someone adds an author that is not already there, it will add it, otherwise it is ignored. Same with genres, and specific combinations of title, author, and genre. Also predictions based off whats in the database as you type.