View book search bar, with params. IfElse on login header
Added new search bar to the book view page, with a dropdown of optional search parameters such as author, genre, and review. Title is searched by default. 
Also added code to the header to display the login link/logged in link accurately based on if you're logged in or not.